# spec/openapi.yaml
openapi: 3.1.0
info:
  title: arw-server
  description: Unified ARW server API surface (headless-first).
  license:
    name: MIT OR Apache-2.0
    identifier: MIT OR Apache-2.0
  version: 0.1.4
paths:
...
  /admin/models:
    get:
      tags:
      - Models
      operationId: models_list_doc
      responses:
        '200':
          description: Models list
          content:
            application/json:
              schema:
                type: array
                items: {}
        '401':
          description: Unauthorized
      description: 'Models: GET /admin/models.'
  /admin/models/add:
    post:
      tags:
      - Models
      operationId: models_add_doc
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelEntry'
        required: true
      responses:
        '200':
          description: Added
          content:
            application/json:
              schema: {}
        '400':
          description: Invalid input
      description: 'Models: POST /admin/models/add.'
  /admin/models/cas_gc:
    post:
      tags:
      - Models
      operationId: models_cas_gc_doc
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CasGcRequest'
        required: true
      responses:
        '200':
          description: GC summary
          content:
            application/json:
              schema: {}
        '501':
          description: CAS GC unavailable
      description: 'Models: POST /admin/models/cas_gc.'
  /admin/models/concurrency:
    get:
      tags:
      - Models
      operationId: models_concurrency_get_doc
      responses:
        '200':
          description: Concurrency
          content:
            application/json:
              schema: {}
      description: 'Models: GET /admin/models/concurrency.'
    post:
      tags:
      - Models
      operationId: models_concurrency_set_doc
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConcurrencyUpdate'
        required: true
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: {}
      description: 'Models: POST /admin/models/concurrency.'
  /admin/models/default:
    get:
      tags:
      - Models
      operationId: models_default_get_doc
      responses:
        '200':
          description: Default model
          content:
            application/json:
              schema: {}
      description: 'Models: GET /admin/models/default.'
    post:
      tags:
      - Models
      operationId: models_default_set_doc
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelId'
        required: true
      responses:
        '200':
          description: Set
          content:
            application/json:
              schema: {}
        '400':
          description: Unknown model
      description: 'Models: POST /admin/models/default.'
  /admin/models/download:
    post:
      tags:
      - Models
      operationId: models_download_doc
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DownloadRequest'
        required: true
      responses:
        '200':
          description: Download accepted
          content:
            application/json:
              schema: {}
        '501':
          description: Download unavailable
      description: 'Models: POST /admin/models/download.'
  /admin/models/download/cancel:
    post:
      tags:
      - Models
      operationId: models_download_cancel_doc
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelId'
        required: true
      responses:
        '200':
          description: Cancellation requested
          content:
            application/json:
              schema: {}
        '501':
          description: Cancel unavailable
      description: 'Models: POST /admin/models/download/cancel.'
...
  /admin/models/load:
    post:
      tags:
      - Models
      operationId: models_load_doc
      responses:
        '200':
          description: Loaded list
          content:
            application/json:
              schema:
                type: array
                items: {}
        '404':
          description: Missing models.json
      description: 'Models: POST /admin/models/load.'
...
  /admin/models/remove:
    post:
      tags:
      - Models
      operationId: models_remove_doc
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelId'
        required: true
      responses:
        '200':
          description: Removed
          content:
            application/json:
              schema: {}
      description: 'Models: POST /admin/models/remove.'
...
  /admin/models/summary:
    get:
      tags:
      - Models
      operationId: models_summary_doc
      responses:
        '200':
          description: Models summary
          content:
            application/json:
              schema: {}
        '401':
          description: Unauthorized
      description: 'Models: GET /admin/models/summary.'
...
  /state/models_hashes:
    get:
      tags:
      - State
      operationId: state_models_hashes_doc
      parameters:
      - name: limit
        in: query
        description: Page size (default 100)
        required: false
        schema:
          type: integer
          minimum: 0
      - name: offset
        in: query
        description: Start offset
        required: false
        schema:
          type: integer
          minimum: 0
      - name: provider
        in: query
        description: Filter by provider
        required: false
        schema:
          type: string
      - name: sort
        in: query
        description: Sort key (bytes|sha256|path|providers_count)
        required: false
        schema:
          type: string
      - name: order
        in: query
        description: Order asc|desc
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Installed hashes
          content:
            application/json:
              schema: {}
      description: 'State: GET /state/models_hashes.'
...
components:
  schemas:
...
    CasGcRequest:
      type: object
      properties:
        ttl_hours:
          type:
          - integer
          - 'null'
          format: int64
          minimum: 0
...
    ConcurrencyUpdate:
      type: object
      properties:
        hard_cap:
          type:
          - integer
          - 'null'
          format: int64
          minimum: 0
        max:
          type:
          - integer
          - 'null'
          format: int64
          minimum: 0
...
    DownloadRequest:
      type: object
      required:
      - id
      - sha256
      properties:
        id:
          type: string
        provider:
          type:
          - string
          - 'null'
        sha256:
          type: string
        url:
          type:
          - string
          - 'null'
...
    ModelEntry:
      type: object
      required:
      - id
      properties:
        id:
          type: string
        path:
          type:
          - string
          - 'null'
        provider:
          type:
          - string
          - 'null'
        sha256:
          type:
          - string
          - 'null'
        status:
          type:
          - string
          - 'null'
    ModelId:
      type: object
      required:
      - id
      properties:
        id:
          type: string
...